+++ Octokit.Reactive/Clients/IObservableRepositoryDeployKeysClient.cs
@@ -33,7 +33,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="number">The id of the deploy key.</param>
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get")]
-        IObservable<DeployKey> Get(int repositoryId, int number);
+        IObservable<DeployKey> Get(long repositoryId, int number);
 
         /// <summary>
         /// Get all deploy keys for a repository.
@@ -52,7 +52,7 @@ namespace Octokit.Reactive
         /// See the <a href="https://developer.github.com/v3/repos/keys/#list"> API documentation</a> for more information.
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository.</param>
-        IObservable<DeployKey> GetAll(int repositoryId);
+        IObservable<DeployKey> GetAll(long repositoryId);
 
         /// <summary>
         /// Get all deploy keys for a repository.
@@ -73,7 +73,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<DeployKey> GetAll(int repositoryId, ApiOptions options);
+        IObservable<DeployKey> GetAll(long repositoryId, ApiOptions options);
 
         /// <summary>
         /// Creates a new deploy key for a repository.
@@ -94,7 +94,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="newDeployKey">The deploy key to create for the repository.</param>
-        IObservable<DeployKey> Create(int repositoryId, NewDeployKey newDeployKey);
+        IObservable<DeployKey> Create(long repositoryId, NewDeployKey newDeployKey);
 
         /// <summary>
         /// Deletes a deploy key from a repository.
@@ -115,6 +115,6 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository.</param>
         /// <param name="number">The id of the deploy key to delete.</param>
-        IObservable<Unit> Delete(int repositoryId, int number);
+        IObservable<Unit> Delete(long repositoryId, int number);
     }
 }
